# src/frontend/staff-portal/src/app/api/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "VTC Staff API",
      "description": "Violation Ticket Centre Staff API",
      "version": "v1"
    },
    "paths": {
...
      "/api/lookup/statutes/v2": {
        "get": {
          "tags": [
            "Lookup"
          ],
          "summary": "Returns a list of Violation Ticket Statutes filtered by given section text (if provided).",
          "parameters": [
            {
              "name": "section",
              "in": "query",
              "description": "Motor vehicle act Section text to query by, ie \"13(1)(a)\" returns \"Motor Vehicle or Trailer without Licence\" contravention, or blank for no filter.",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Statute"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Request lacks valid authentication credentials.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/lookup/languages/v2": {
        "get": {
          "tags": [
            "Lookup"
          ],
          "summary": "Returns a list of Languages.",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Language"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Request lacks valid authentication credentials.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/lookup/agencies/v2": {
        "get": {
          "tags": [
            "Lookup"
          ],
          "summary": "Returns a list of agencies.",
          "parameters": [
            {
              "name": "type",
              "in": "query",
              "description": "The agency type code",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Agency"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Request lacks valid authentication credentials.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/lookup/provinces/v2": {
        "get": {
          "tags": [
            "Lookup"
          ],
          "summary": "Returns a list of provinces.",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Province"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Request lacks valid authentication credentials.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/lookup/countries/v2": {
        "get": {
          "tags": [
            "Lookup"
          ],
          "summary": "Returns a list of countries.",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Country"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Request lacks valid authentication credentials.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
      "/api/lookup/dispute-case-file-statuses/v2": {
        "get": {
          "tags": [
            "Lookup"
          ],
          "summary": "Returns a list of dispute case file status types.",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DisputeCaseFileStatus"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Request lacks valid authentication credentials.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Agency": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "nullable": true
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "typeCode": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "Country": {
          "type": "object",
          "properties": {
            "ctryId": {
              "type": "string",
              "nullable": true
            },
            "ctryLongNm": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "DisputeCaseFileStatus": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "Language": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "ProblemDetails": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "nullable": true
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "status": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "detail": {
              "type": "string",
              "nullable": true
            },
            "instance": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": { }
        },
        "Province": {
          "type": "object",
          "properties": {
            "ctryId": {
              "type": "string",
              "nullable": true
            },
            "provSeqNo": {
              "type": "string",
              "nullable": true
            },
            "provNm": {
              "type": "string",
              "nullable": true
            },
            "provAbbreviationCd": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "Statute": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "nullable": true
            },
            "actCode": {
              "type": "string",
              "nullable": true
            },
            "sectionText": {
              "type": "string",
              "nullable": true
            },
            "subsectionText": {
              "type": "string",
              "nullable": true
            },
            "paragraphText": {
              "type": "string",
              "nullable": true
            },
            "subparagraphText": {
              "type": "string",
              "nullable": true
            },
            "code": {
              "type": "string",
              "nullable": true
            },
            "shortDescriptionText": {
              "type": "string",
              "nullable": true
            },
            "descriptionText": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
      },
      "securitySchemes": {
        "Bearer": {
          "type": "apiKey",
          "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
          "name": "Authorization",
          "in": "header"
        }
      }
    },
    "security": [
      {
        "Bearer": [ ]
      }
    ]
  }